add conditional to delete file locally (not using AWS deletion yet), change constant names to be more semtantically accurate

+++ DocumentSummarizer/src/main/java/com/jakeporter/DocumentSummarizer/service/FileService.java
@@ -1,11 +1,8 @@
 package com.jakeporter.DocumentSummarizer.service;
 
-import com.jakeporter.DocumentSummarizer.exceptions.FileDeletionException;
 import com.jakeporter.DocumentSummarizer.exceptions.GenericFileException;
 import com.jakeporter.DocumentSummarizer.utilities.fileUtils.FileInfoGetter;
 import com.jakeporter.DocumentSummarizer.utilities.fileUtils.removers.FileRemover;
-import com.jakeporter.DocumentSummarizer.utilities.fileUtils.removers.LocalRemover;
-import com.jakeporter.DocumentSummarizer.utilities.fileUtils.uploaders.AWSS3Uploader;
 import com.jakeporter.DocumentSummarizer.utilities.fileUtils.uploaders.FileUploader;
 import com.jakeporter.DocumentSummarizer.utilities.summarizers.DocumentSummarizer;
 import com.jakeporter.DocumentSummarizer.utilities.summarizers.PythonSummarizer;
@@ -24,18 +21,19 @@ import java.util.Set;
 @Service
 public class FileService {
 
-    private final static boolean IS_LOCAL = false;
-    private final static String BUILD = IS_LOCAL ? "local" : "aws";
+    private final static boolean IS_PERSISTING_LOCALLY = false;
+    private final static String UPLOADER = IS_PERSISTING_LOCALLY ? "localUploader" : "awsUploader";
+    private final static String REMOVER = IS_PERSISTING_LOCALLY ? "localRemover" : "awsRemover";
 
     @Value("${upload.dir:${user.home}}")
     private String UPLOAD_DIRECTORY;
 
     @Autowired
-    @Qualifier(BUILD)
+    @Qualifier(UPLOADER)
     private FileUploader uploader;
 
     @Autowired
-    @Qualifier(BUILD)
+    @Qualifier(REMOVER)
     private FileRemover remover;
 
     Logger logger = LoggerFactory.getLogger(this.getClass());
@@ -46,7 +44,7 @@ public class FileService {
 
     public Set<String> summarize(MultipartFile file) {
         FileInfoGetter infoGetter = new FileInfoGetter();
-        if (!IS_LOCAL) {
+        if (!IS_PERSISTING_LOCALLY) {
             file = getFileFromS3(file);
         }
         FileType fileType = infoGetter.getFileType(FilenameUtils.getExtension(file.getOriginalFilename()));
@@ -59,7 +57,9 @@ public class FileService {
         } catch (Exception e) { // Don't know what uncaught exceptions could throw this at the moment, but I want to ensure the deletion of any uploaded files
             throw new GenericFileException("Something went wrong.");
         } finally {
-            remover.deleteFile(infoGetter.getFileLocation(file, UPLOAD_DIRECTORY));
+            if (IS_PERSISTING_LOCALLY) {
+                remover.deleteFile(infoGetter.getFileLocation(file, UPLOAD_DIRECTORY));
+            }
         }
         return summaries;
     }
